+++ handler.go
@@ -1,10 +1,12 @@
 package main
 
 import (
+	"context"
 	"net/http"
 	"net/http/httputil"
 
 	"github.com/mia-platform/glogger/v2"
+	"github.com/open-policy-agent/opa/rego"
 )
 
 const URL_SCHEME = "http"
@@ -13,9 +15,39 @@ func rbacHandler(w http.ResponseWriter, req *http.Request) {
 	env, err := GetEnv(req.Context())
 	if err != nil {
 		glogger.Get(req.Context()).WithError(err).Error("no env found in context")
-		w.WriteHeader(http.StatusInternalServerError)
+		failResponse(w, "no environment found in context")
 		return
 	}
+
+	opaEvaluator, err := GetOPAEvaluator(req.Context())
+	if err != nil {
+		glogger.Get(req.Context()).WithError(err).Error("no policy evaluator found in context")
+		failResponse(w, "no policy evaluator found in context")
+		return
+	}
+
+	input := map[string]interface{}{
+		"request": map[string]interface{}{
+			"method":  req.Method,
+			"path":    req.URL.Path,
+			"headers": req.Header,
+			"query":   req.URL.Query(),
+		},
+	}
+
+	results, err := opaEvaluator.PermissionQuery.Eval(context.TODO(), rego.EvalInput(input))
+	if err != nil {
+		glogger.Get(req.Context()).WithError(err).Error("policy eval failed")
+		failResponse(w, "policy eval failed")
+		return
+	}
+
+	if !results.Allowed() {
+		glogger.Get(req.Context()).WithError(err).Error("policy resulted in not allowed")
+		failResponseWithCode(w, http.StatusForbidden, "RBAC policy evaluation failed")
+		return
+	}
+
 	targetHostFromEnv := env.TargetServiceHost
 	proxy := httputil.ReverseProxy{
 		Director: func(req *http.Request) {